pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.2"
lxml = "^5.1.0"
faker = "^24.4.0"
kafka-python = "^2.0.2"
pillow = "^10.2.0"
python-crontab = "^3.0.0"
pymongo = "^4.6.3"
black = "^24.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
